+++ pydit/__init__.py
@@ -8,3 +8,5 @@ from .filemanager import FileManager
 from .logger import setup_logging
 from .logger import start_logging_info
 from .logger import start_logging_debug
+
+__version__ = "0.0.7"